DATAJPA-507 - LocakModeRepositoryPostProcessor now caches lock metadata.
We now cache the LockModeType annotation found for a method to avoid repeated lookups of the annotation using AnnotationUtils.findAnnotation(…) as the call to it is quite expensive (as it needs to travers the entire method hierarchy and inspect meta-annotations etc.).

This is a partial back-port of the fix for DATAJPA-173 with only the performance improvement, leaving the functional improvements in the master branch.