DATAJPA-519 - Spec compliant mapping file URI handling in ClasspathScanningPersistenceUnitPostProcessor.
JPA requires mapping file locations to be classpath relative. Previously we handed the full path to the file (including protocol and JAR path) to the provider, which they happened to understand by accident. To be more spec compliant, we now massage the detected URIs into the expected format.

Original pull request: #81.