DATAJPA-551 - Add support for limiting the query result in the query derivation mechanism.
We now support to restrict the results of a query method via the "Top" or "First" keyword within the method name. We adjust the firstResult offset of the query in case we detect a tighter bound on the result size limit form the derived query method that is smaller than the page size of the given Pageable.

Added test cases for Paged and Sliced query executions.

Original pull request: #96.