DATAJPA-858 - Fixed collection contains handling for nested path traversals.
Previously the first property was checked for being a collection to trigger collection contains handling. This is wrong for nested property traversals as they might end up in a String for which a like binding has to be applied then.

We're now inspecting the leaf property for being a collection to trigger that special binding.