DATAJPA-1174 - Reinstantiated null handling in Jsr310JpaConverters.
Jsr310Converters in Spring Data Commons has been aligned with Spring's general interface contract for parameters (expecting them not to be called with null values) which breaks the JPA AttributeConverter implementations as they have to deal with null values.

We now added the handling previously done in Jsr310Converters to Jsr310JpaConverters.