DATAJPA-949 - Added support for JPA 2.2-native stream execution.
We're now discovering JPA 2.2 on the classpath and invoke Query.getResultStream() directly in favor of our own implementation based on individual persistence provider APIs. We also use the method it it is already declared on their JPA 2.1 Query implementation. That allows us to use e.g. Hibernate 5.2's implementation of it, even if it is still compiled against JPA 2.1.