DATAJPA-1209 - Re-instantiate compatibility with Hibernate < 5.2.11 for projection on native queries.
We now explicitly check for the presence of Hibernate 5.2.11 or better for native query projections using Tuple during query execution. Doing so has the advantage of properly producing a mapped result so that projection method invocations can use the map keys to route the calls to the right value.

With versions before 5.2.11 we rely on the select expression listing the columns to return in the order matching the declaration order of the methods in the projection interface (which is theoretically unstable die to the reflection API not guaranteeing a stable order).

That said, it's strongly recommended to use both Hibernate in 5.2.11 or better in combination with explicitly aliased columns in select clauses for native queries using projections.

Related ticket: DATAJPA-980.