DATAJPA-1241 - Removed dynamic check for getResultStream() on JPA query implementation classes.
We removed the dynamic, one-time lookup of a getResultStream() method on a JPA Query implementation which was basically introduced to accommodate the fact, that Hibernate 5.2's implementation already ships the method, but doesn't actually implement JPA 2.2. However, in varying circumstances, Hibernate returns a JPA Query proxy for the call to EntityManager.getQuery(…). If that's the case we can't implement the detected implementation method on that instance. Even worse, depending on which of the two we see first (Hibernate's implementation class or the JPA Query proxy) the optimization might not even kick in as the detection of the method is a one-time effort.

Original pull request: #241.