DATAJPA-1280 - Tweaked the type handling for named native queries.
We now only use Tuple as target type to execute a named native query if a projection interface, i.e. an interface that's not part of the entity type's type hierarchy is used as return type. If a type within the entity types hierarchy is returned we use the returned type as type hint. In all other cases (usually DTOs) we rely on the named native query being backed by a custom JPA 2.1 SQL result set mapping.