DATAJPA-1519 - Wildcard escaping in LIKE-queries.
Derived queries with `contains`, `startWith` or `endsWith` predicates now use wildcard escaping in order to avoid selecting more then requested. A function `escape(String, String)` is made available in SpEL expressions as well for the same purpose.

Note that annotated queries and derived queries with LIKE conditions don't get escaped because it is not clear if the use of wildcards in the argument is intended or not. Improved reference documentation accordingly.