DATAJPA-1535 - Additional test for entities with assigned identifiers.
Added test case for the new optimized way of handling deletes of non persisted instances which use manually assigned identifiers. This test originally broke as manually marking the entity as new was done in a @PostPersist callback. This needed to be changed to happen in @PrePersist as the actual persistence operation might not have been flushed and the entity might return in new state from CrudRepository.save(…).