Disable parameter names for test compilation.
We now no longer use compiled parameter names for tests to avoid interference with Eclipselink's mechanism of using these improperly for stored procedure calls. Especially the fact that a parameter is named "in" lets the database consider using an IN clause instead of the parameter name.

When configuring the compile plugin, we also need to ensure that the AspectJ plugin post-processes the compiled classes and doesn't run before the Maven compiler plugin.

Closes: #2298
See also https://github.com/spring-projects/spring-data-build/issues/1513